requirements.txt
bcrypt==3.2.0
cffi==1.15.0
cryptography==35.0.0
future==0.18.2
netmiko==3.4.0
ntc-templates==3.0.0
paramiko==2.8.0
pycparser==2.21
PyNaCl==1.4.0
pyserial==3.5
PyYAML==6.0
scp==0.14.1
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
